Migrate wiki commands to new structue
